move card component into seperate file, add icons
